+++ src/MoBi.Assets/MoBi.Assets.csproj
@@ -26,9 +26,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Assets" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Assets.Images" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Core" Version="12.0.0-ojEDY" />
+    <PackageReference Include="OSPSuite.Assets" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Assets.Images" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Core" Version="12.0.188" />
   </ItemGroup>
 
   <ItemGroup>
