+++ src/MoBi.Core/Commands/EditParameterValuePathCommand.cs
@@ -6,7 +6,7 @@ using OSPSuite.Core.Domain.Builder;
 
 namespace MoBi.Core.Commands
 {
-   public class EditParameterValuePathCommand : EditPathAndValueEntityPathCommand<IBuildingBlock<ParameterValue>, ParameterValue>
+   public class EditParameterValuePathCommand : EditPathAndValueEntityPathCommand<ILookupBuildingBlock<ParameterValue>, ParameterValue>
    {
       /// <summary>
       ///    Changes a path for a Parameter value
@@ -14,14 +14,14 @@ namespace MoBi.Core.Commands
       /// <param name="buildingBlock">The building block this start value is part of</param>
       /// <param name="parameterValue">The parameter value being modified</param>
       /// <param name="newContainerPath">The new container path for the start value</param>
-      public EditParameterValuePathCommand(IBuildingBlock<ParameterValue> buildingBlock, ParameterValue parameterValue, ObjectPath newContainerPath)
+      public EditParameterValuePathCommand(ILookupBuildingBlock<ParameterValue> buildingBlock, ParameterValue parameterValue, ObjectPath newContainerPath)
          : base(buildingBlock, parameterValue, newContainerPath)
       {
       }
 
       protected override ICommand<IMoBiContext> GetInverseCommand(IMoBiContext context)
       {
-         return new EditParameterValuePathCommand(_buildingBlock, _buildingBlock.Single(x => Equals(x.Path, _newPath)), _originalContainerPath).AsInverseFor(this);
+         return new EditParameterValuePathCommand(_buildingBlock, _buildingBlock.ByPath(_newPath), _originalContainerPath).AsInverseFor(this);
       }
    }
 }
