+++ src/MoBi.Core/Commands/RemovePathAndValueEntityFromBuildingBlockCommand.cs
@@ -9,16 +9,16 @@ using OSPSuite.Core.Domain.Builder;
 
 namespace MoBi.Core.Commands
 {
-   public class RemovePathAndValueEntityFromBuildingBlockCommand<T> : BuildingBlockChangeCommandBase<IBuildingBlock<T>> where T : PathAndValueEntity
+   public class RemovePathAndValueEntityFromBuildingBlockCommand<T> : BuildingBlockChangeCommandBase<ILookupBuildingBlock<T>> where T : PathAndValueEntity
    {
       private readonly T _originalEntity;
 
-      public RemovePathAndValueEntityFromBuildingBlockCommand(IBuildingBlock<T> parent, ObjectPath path) : base(parent)
+      public RemovePathAndValueEntityFromBuildingBlockCommand(ILookupBuildingBlock<T> parent, ObjectPath path) : base(parent)
       {
          CommandType = AppConstants.Commands.DeleteCommand;
 
          ObjectType = new ObjectTypeResolver().TypeFor<T>();
-         _originalEntity = _buildingBlock.Single(x => Equals(x.Path, path));
+         _originalEntity = _buildingBlock.ByPath(path);
          Description = AppConstants.Commands.RemovePathAndValueEntity(_originalEntity, parent.Name, ObjectType);
       }
 
@@ -43,7 +43,7 @@ namespace MoBi.Core.Commands
 
    public class RemoveInitialConditionFromBuildingBlockCommand : RemovePathAndValueEntityFromBuildingBlockCommand<InitialCondition>
    {
-      public RemoveInitialConditionFromBuildingBlockCommand(IBuildingBlock<InitialCondition> parent, ObjectPath path)
+      public RemoveInitialConditionFromBuildingBlockCommand(ILookupBuildingBlock<InitialCondition> parent, ObjectPath path)
          : base(parent, path)
       {
       }
