+++ src/MoBi.Core/MoBi.Core.csproj
@@ -33,13 +33,13 @@
     <PackageReference Include="FluentNHibernate" Version="2.1.2" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="OSPSuite.Utility" Version="4.0.0.4" />
-    <PackageReference Include="OSPSuite.Core" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Assets" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Assets.Images" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Export" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Reporting" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Serialization" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="12.0.0-ojEDY" />
+    <PackageReference Include="OSPSuite.Core" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Assets" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Assets.Images" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Export" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Reporting" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Serialization" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="12.0.188" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.7.0" />
   </ItemGroup>
