+++ src/MoBi.Engine/MoBi.Engine.csproj
@@ -32,8 +32,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Core" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Assets" Version="12.0.0-ojEDY" />
+    <PackageReference Include="OSPSuite.Core" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Assets" Version="12.0.188" />
   </ItemGroup>
 
   <ItemGroup>
