+++ src/MoBi.Presentation/Tasks/Interaction/AbstractStartValuePathTask.cs
@@ -9,7 +9,7 @@ using OSPSuite.Utility.Extensions;
 
 namespace MoBi.Presentation.Tasks.Interaction
 {
-   public interface IStartValuePathTask<TBuildingBlock, TPathAndValueEntity> where TBuildingBlock : IBuildingBlock<TPathAndValueEntity>, IBuildingBlock where TPathAndValueEntity : PathAndValueEntity
+   public interface IStartValuePathTask<TBuildingBlock, TPathAndValueEntity> where TBuildingBlock : ILookupBuildingBlock<TPathAndValueEntity>, IBuildingBlock where TPathAndValueEntity : PathAndValueEntity
    {
       /// <summary>
       ///    Updates the name of the <paramref name="pathAndValueEntity" /> to <paramref name="newName" /> and returns the
@@ -32,7 +32,7 @@ namespace MoBi.Presentation.Tasks.Interaction
       ///    returns the non executed command
       /// </summary>
       /// <returns>The command corresponding to the rename of the start value. The command was not run yet</returns>
-      IMoBiCommand UpdateNameCommand(IBuildingBlock<TPathAndValueEntity> startValues, TPathAndValueEntity pathAndValueEntity, string newName);
+      IMoBiCommand UpdateNameCommand(ILookupBuildingBlock<TPathAndValueEntity> startValues, TPathAndValueEntity pathAndValueEntity, string newName);
 
       /// <summary>
       ///    Creates a command corresponding to the update of the path entry at <paramref name="indexToUpdate" /> of the
@@ -42,7 +42,7 @@ namespace MoBi.Presentation.Tasks.Interaction
       ///    The command corresponding to the path update at <paramref name="indexToUpdate" /> of the start value. The
       ///    command was not run yet
       /// </returns>
-      IMoBiCommand UpdateContainerPathCommand(IBuildingBlock<TPathAndValueEntity> buildingBlock, TPathAndValueEntity pathAndValueEntity, int indexToUpdate, string newValue);
+      IMoBiCommand UpdateContainerPathCommand(ILookupBuildingBlock<TPathAndValueEntity> buildingBlock, TPathAndValueEntity pathAndValueEntity, int indexToUpdate, string newValue);
 
       /// <summary>
       ///    Checks that the formula is equivalent for the start value. This includes evaluation of constant formula to a double
@@ -53,7 +53,7 @@ namespace MoBi.Presentation.Tasks.Interaction
       bool HasEquivalentFormula(PathAndValueEntity pathAndValueEntity, IFormula targetFormula);
    }
 
-   public abstract class AbstractStartValuePathTask<TBuildingBlock, TPathAndValueEntity> : IStartValuePathTask<TBuildingBlock, TPathAndValueEntity> where TBuildingBlock : IBuildingBlock<TPathAndValueEntity> where TPathAndValueEntity : PathAndValueEntity
+   public abstract class AbstractStartValuePathTask<TBuildingBlock, TPathAndValueEntity> : IStartValuePathTask<TBuildingBlock, TPathAndValueEntity> where TBuildingBlock : ILookupBuildingBlock<TPathAndValueEntity> where TPathAndValueEntity : PathAndValueEntity
    {
       private readonly IFormulaTask _formulaTask;
       private readonly IMoBiContext _context;
@@ -74,8 +74,8 @@ namespace MoBi.Presentation.Tasks.Interaction
          return UpdateContainerPathCommand(startValues, pathAndValueEntity, indexToUpdate, newValue).Run(_context);
       }
 
-      public abstract IMoBiCommand UpdateNameCommand(IBuildingBlock<TPathAndValueEntity> startValues, TPathAndValueEntity pathAndValueEntity, string newName);
-      public abstract IMoBiCommand UpdateContainerPathCommand(IBuildingBlock<TPathAndValueEntity> buildingBlock, TPathAndValueEntity pathAndValueEntity, int indexToUpdate, string newValue);
+      public abstract IMoBiCommand UpdateNameCommand(ILookupBuildingBlock<TPathAndValueEntity> startValues, TPathAndValueEntity pathAndValueEntity, string newName);
+      public abstract IMoBiCommand UpdateContainerPathCommand(ILookupBuildingBlock<TPathAndValueEntity> buildingBlock, TPathAndValueEntity pathAndValueEntity, int indexToUpdate, string newValue);
 
       public static void ConfigureTargetPath(int indexToUpdate, string newValue, ObjectPath targetPath)
       {
