+++ src/MoBi.Presentation/Tasks/Interaction/InitialConditionPathTask.cs
@@ -6,22 +6,22 @@ using OSPSuite.Core.Domain.Services;
 
 namespace MoBi.Presentation.Tasks.Interaction
 {
-   public interface IInitialConditionPathTask : IStartValuePathTask<IBuildingBlock<InitialCondition>, InitialCondition>
+   public interface IInitialConditionPathTask : IStartValuePathTask<ILookupBuildingBlock<InitialCondition>, InitialCondition>
    {
    }
 
-   public class InitialConditionPathTask : AbstractStartValuePathTask<IBuildingBlock<InitialCondition>, InitialCondition>, IInitialConditionPathTask
+   public class InitialConditionPathTask : AbstractStartValuePathTask<ILookupBuildingBlock<InitialCondition>, InitialCondition>, IInitialConditionPathTask
    {
       public InitialConditionPathTask(IFormulaTask formulaTask, IMoBiContext context) : base(formulaTask, context)
       {
       }
 
-      public override IMoBiCommand UpdateNameCommand(IBuildingBlock<InitialCondition> startValues, InitialCondition pathAndValueEntity, string newValue)
+      public override IMoBiCommand UpdateNameCommand(ILookupBuildingBlock<InitialCondition> startValues, InitialCondition pathAndValueEntity, string newValue)
       {
          return new ChangeInitialConditionNameCommand(startValues, pathAndValueEntity.Path, newValue);
       }
 
-      public override IMoBiCommand UpdateContainerPathCommand(IBuildingBlock<InitialCondition> buildingBlock, InitialCondition pathAndValueEntity, int indexToUpdate, string newValue)
+      public override IMoBiCommand UpdateContainerPathCommand(ILookupBuildingBlock<InitialCondition> buildingBlock, InitialCondition pathAndValueEntity, int indexToUpdate, string newValue)
       {
          var targetPath = pathAndValueEntity.ContainerPath.Clone<ObjectPath>();
