+++ src/MoBi.Presentation/Tasks/Interaction/ParameterStartValuesPathTask.cs
@@ -6,22 +6,22 @@ using OSPSuite.Core.Domain.Services;
 
 namespace MoBi.Presentation.Tasks.Interaction
 {
-   public interface IParameterValuePathTask : IStartValuePathTask<IBuildingBlock<ParameterValue>, ParameterValue>
+   public interface IParameterValuePathTask : IStartValuePathTask<ILookupBuildingBlock<ParameterValue>, ParameterValue>
    {
    }
 
-   public class ParameterValuePathTask : AbstractStartValuePathTask<IBuildingBlock<ParameterValue>, ParameterValue>, IParameterValuePathTask
+   public class ParameterValuePathTask : AbstractStartValuePathTask<ILookupBuildingBlock<ParameterValue>, ParameterValue>, IParameterValuePathTask
    {
       public ParameterValuePathTask(IFormulaTask formulaTask, IMoBiContext context) : base(formulaTask, context)
       {
       }
 
-      public override IMoBiCommand UpdateNameCommand(IBuildingBlock<ParameterValue> startValues, ParameterValue pathAndValueEntity, string newValue)
+      public override IMoBiCommand UpdateNameCommand(ILookupBuildingBlock<ParameterValue> startValues, ParameterValue pathAndValueEntity, string newValue)
       {
          return new ChangeParameterValueNameCommand(startValues, pathAndValueEntity.Path, newValue);
       }
 
-      public override IMoBiCommand UpdateContainerPathCommand(IBuildingBlock<ParameterValue> buildingBlock, ParameterValue pathAndValueEntity, int indexToUpdate, string newValue)
+      public override IMoBiCommand UpdateContainerPathCommand(ILookupBuildingBlock<ParameterValue> buildingBlock, ParameterValue pathAndValueEntity, int indexToUpdate, string newValue)
       {
          var targetPath = pathAndValueEntity.ContainerPath.Clone<ObjectPath>();
          if (indexToUpdate > targetPath.Count)
