+++ src/MoBi.Presentation/UICommand/InitialConditionsExtendUICommand.cs
@@ -6,7 +6,7 @@ namespace MoBi.Presentation.UICommand
 {
    internal class InitialConditionsExtendUICommand : AbstractStartValueSubjectRetrieverUICommand<InitialConditionsBuildingBlock, InitialCondition>
    {
-      public InitialConditionsExtendUICommand(IInitialConditionsTask startValueTasks, IActiveSubjectRetriever activeSubjectRetriever)
+      public InitialConditionsExtendUICommand(IInitialConditionsTask<InitialConditionsBuildingBlock> startValueTasks, IActiveSubjectRetriever activeSubjectRetriever)
          : base(startValueTasks, activeSubjectRetriever)
       {
       }
