+++ src/MoBi.Presentation/Views/IInitialConditionsView.cs
@@ -1,22 +1,13 @@
 using MoBi.Presentation.DTO;
 using MoBi.Presentation.Presenter;
-using OSPSuite.Presentation.Presenters;
 using OSPSuite.Presentation.Views;
 
 namespace MoBi.Presentation.Views
 {
-   public interface IBuildingBlockWithInitialConditionsView<TPresenter> : IView<TPresenter>, IPathAndValueEntitiesView<InitialConditionDTO> where TPresenter : IPresenter
+   public interface IInitialConditionsView : IView<IBuildingBlockWithInitialConditionsPresenter>, IPathAndValueEntitiesView<InitialConditionDTO>
    {
       void AddIsPresentSelectionView(IView view);
 
       void AddNegativeValuesAllowedSelectionView(IView view);
-   }
-
-   public interface IInitialConditionsView : IBuildingBlockWithInitialConditionsView<IInitialConditionsPresenter>
-   {
-   }
-
-   public interface IExpressionProfileInitialConditionsView : IBuildingBlockWithInitialConditionsView<IExpressionProfileInitialConditionsPresenter>
-   {
    }
 }
