+++ src/MoBi.UI/MoBi.UI.csproj
@@ -27,11 +27,11 @@
     <PackageReference Include="OSPSuite.DataBinding.DevExpress" Version="6.0.0.2" />
     <PackageReference Include="OSPSuite.DevExpress" Version="21.2.3" />
     <PackageReference Include="OSPSuite.Utility" Version="4.0.0.4" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Presentation" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.UI" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Core" Version="12.0.0-ojEDY" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Presentation" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.UI" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Core" Version="12.0.188" />
    </ItemGroup>
   <ItemGroup>
     <None Include="..\..\LICENSE">
