+++ src/MoBi.UI/Views/InitialConditionsView.resx
