+++ tests/MoBi.Tests/MoBi.Tests.csproj
@@ -45,10 +45,10 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
     <PackageReference Include="nunit" Version="3.13.2" />
     <PackageReference Include="OSPSuite.BDDHelper" Version="4.0.0.1" />
-    <PackageReference Include="OSPSuite.Core" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="12.0.0-ojEDY" />
-    <PackageReference Include="OSPSuite.UI" Version="12.0.0-ojEDY" />
+    <PackageReference Include="OSPSuite.Core" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="12.0.188" />
+    <PackageReference Include="OSPSuite.UI" Version="12.0.188" />
     <PackageReference Include="OSPSuite.FuncParser" Version="4.0.0.54" GeneratePathProperty="true" />
     <PackageReference Include="OSPSuite.SimModel" Version="4.0.0.53" GeneratePathProperty="true" />
     <PackageReference Include="OSPSuite.SimModelSolver_CVODES" Version="4.1.0.8" GeneratePathProperty="true" />
